Create golfCode.js
This is a simple Javascript game from Freecodecamp.org and all rights are reserved to them.
Following the description of the problem:  

In the game of golf each hole has a par meaning the average number of strokes a golfer is expected to make in order to sink the ball in a hole to complete the play. Depending on how far above or below par your strokes are, there is a different nickname.

Your function will be passed par and strokes arguments. Return the correct string according to this table which lists the strokes in order of priority; top (highest) to bottom (lowest):

Strokes	                                     Return
1                                              	"Hole-in-one!"
<= par - 2	                                "Eagle"
par - 1                                        	"Birdie"
par	                                                "Par"
par + 1	                                        "Bogey"
par + 2                                        	"Double Bogey"
>= par + 3                                	"Go Home!"


NOTE : par and strokes will always be numeric and positive. We have added an array of all the names for your convenience.